purple-air/requirements.txt
requests
python-aqi
paho.mqtt
